api-gateway-mobile/requirements.txt
Flask==2.3.2
protobuf>=4.21.0
grpcio>=1.54.0
grpcio-tools==1.64.0
requests==2.31.0
flask-cors
